# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API for comparing frogs and other things",
        "title": "Frog Smash API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Logs in a user with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Logs out a user by clearing the refresh token cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User logout",
                "responses": {}
            }
        },
        "/auth/me": {
            "get": {
                "description": "Retrieves the currently authenticated user's information",
                "produces": [
                    "application/json"
                ],
                "summary": "Get current user",
                "responses": {}
            }
        },
        "/auth/refresh-token": {
            "post": {
                "description": "Refreshes the JWT token using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Refresh JWT token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Registers a new user with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegistrationRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/comparison/compare": {
            "post": {
                "description": "Records the result of a comparison between two items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Compare two items",
                "parameters": [
                    {
                        "description": "Comparison Request",
                        "name": "compareRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CompareRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/comparison/items": {
            "get": {
                "description": "Retrieves two distinct items for comparison",
                "produces": [
                    "application/json"
                ],
                "summary": "Get two items for comparison",
                "responses": {}
            }
        },
        "/comparison/latest-submission": {
            "get": {
                "description": "Retrieves the time of the latest submission by the user",
                "produces": [
                    "application/json"
                ],
                "summary": "Get time of latest submission",
                "responses": {}
            }
        },
        "/comparison/leaderboard": {
            "get": {
                "description": "Retrieves a paginated leaderboard of items",
                "produces": [
                    "application/json"
                ],
                "summary": "Get leaderboard",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/comparison/submit-contender": {
            "post": {
                "description": "Uploads an image to the server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload an image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/verify": {
            "post": {
                "description": "Verifies the user's email using a verification code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify user email",
                "parameters": [
                    {
                        "description": "Verification code payload",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VerificationRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/verify/resend-email-anonymous": {
            "post": {
                "description": "Resends the verification email to the user without authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Resend verification email (anonymous)",
                "responses": {}
            }
        },
        "/verify/resend-verification": {
            "post": {
                "description": "Resends the verification email to the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Resend verification email",
                "responses": {}
            }
        }
    },
    "definitions": {
        "dto.CompareRequest": {
            "description": "Request payload for comparing two items",
            "type": "object",
            "properties": {
                "loser_id": {
                    "type": "string"
                },
                "winner_id": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginRequest": {
            "description": "Request payload for user login",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginResponse": {
            "description": "Response payload for user login",
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.UserResponse"
                }
            }
        },
        "dto.UserRegistrationRequest": {
            "description": "Request payload for user registration",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 12
                }
            }
        },
        "dto.UserResponse": {
            "description": "User data returned with login response",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isVerified": {
                    "type": "boolean"
                }
            }
        },
        "dto.VerificationRequest": {
            "description": "Request payload for user verification",
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        }
    }
}